websocket: separate connection logic from websocket
